feat(layout): adjust flex-layout to optimize display of map
