+++ src/index.css
@@ -14,11 +14,6 @@
   }
 }
 
-.leaflet-container {
-  height: 100vh;
-  width: 90vw;
-}
-
 #controlButton {
   display: flex;
   align-items: center;
